requirements.txt
# Core ML frameworks
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
sentence-transformers>=2.2.0

# Parameter-efficient fine-tuning
peft>=0.6.0
trl>=0.7.0
bitsandbytes>=0.41.0

# Vector search
faiss-gpu>=1.7.0  # Use faiss-cpu if no GPU available

# Evaluation metrics
evaluate>=0.4.0
nltk>=3.8.0
rouge-score>=0.1.2
bert-score>=0.3.13

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Experiment tracking
wandb>=0.15.0

# NLP
spacy>=3.6.0

# Progress bars
tqdm>=4.65.0

# Environment variable loading
python-dotenv>=1.0.0


# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
